refactor: Create UserFunds component to extract this logic from myfunds page component
